`long-lines-render': highlight long part
`long-lines-render': highlight the long part with
`long-lines-render-highlight-face'.

Call `long-lines-goto-column' to compute the
long column in that function separately and not in `long-lines-in-buffer', since
`current-column' is much faster and long lines are usually rare.

`long-lines-render' doesn't save `point', since calling `save-excursion' every
time has some overhead. `save-excursion' is called around any calls to it.